requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-mysql==4.14.0
django-seed==0.3.1
django-seeding==1.3.5
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==30.1.0
inflection==0.5.1
Markdown==3.7
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
uritemplate==4.1.1
